requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.10.0=py36h9f0ad1d_0
astor=0.8.1=pyh9f0ad1d_0
audioread=2.1.8=py36h9f0ad1d_2
brotlipy=0.7.0=py36h8c4c3a4_1000
bzip2=1.0.8=h516909a_3
c-ares=1.16.1=h516909a_3
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py36h9f0ad1d_0
cffi=1.11.5=py36_0
chardet=3.0.4=py36h9f0ad1d_1007
cryptography=3.1=py36h45558ae_0
cycler=0.10.0=py_2
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
expat=2.2.9=he6710b0_2
ffmpeg=4.3.1=h167e202_0
ffmpeg-python=0.2.0=py_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.2=h5ab3b9f_0
future=0.18.2=py36h9f0ad1d_1
gast=0.2.2=py_0
gettext=0.19.8.1=h5e8e0c9_1
glib=2.65.0=h3eb4bd4_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.31.0=py36h769ab6c_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.10.0=nompi_py36hecadee3_104
hdf5=1.10.6=nompi_h3c11f04_101
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=1.7.0=py36h9f0ad1d_0
importlib_metadata=1.7.0=0
importlib_resources=3.0.0=py36h9f0ad1d_0
joblib=0.16.0=py_0
jpeg=9b=h024ee3a_2
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_0
kiwisolver=1.2.0=py36hdb11119_0
lame=3.100=h14c3975_1001
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=17_openblas
libcblas=3.8.0=17_openblas
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libflac=1.3.3=he1b5a44_0
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=hdf63c60_16
libiconv=1.16=h516909a_0
liblapack=3.8.0=17_openblas
libllvm8=8.0.1=hc9558a2_0
libogg=1.3.2=h516909a_1002
libopenblas=0.3.10=pthreads_hb3c22a3_4
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0=h8b12597_0
librosa=0.7.2=py_1
libsndfile=1.0.29=he1b5a44_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libvorbis=1.3.7=he1b5a44_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
llvmlite=0.31.0=py36hfa65bc7_1
markdown=3.2.2=py_0
matplotlib-base=3.1.1=py36hfd891ef_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=h1bed415_1002
norbert=0.2.1=py_0
numba=0.48.0=py36hb3f55d8_0
numpy=1.19.1=py36h3849536_2
openh264=2.1.1=h8b12597_0
openssl=1.1.1g=h516909a_1
opt_einsum=3.3.0=py_0
pandas=0.25.1=py36hb3f55d8_0
pcre=8.44=he6710b0_0
pip=20.2.2=py36_0
protobuf=3.13.0=py36h831f99a_0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py36h05f1152_2
pysocks=1.7.1=py36h9f0ad1d_1
pysoundfile=0.10.2=py_1001
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytz=2020.1=pyh9f0ad1d_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
requests=2.24.0=pyh9f0ad1d_0
resampy=0.2.2=py_0
scikit-learn=0.23.2=py36hfb379a7_0
scipy=1.5.2=py36h3a855aa_0
setuptools=49.6.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=pyh9f0ad1d_0
spleeter=1.5.3=py36h9f0ad1d_2
sqlite=3.33.0=h62c20be_0
tensorboard=1.15.0=py36_0
tensorflow=1.15.0=mkl_py36h4920b83_0
tensorflow-base=1.15.0=mkl_py36he1670d9_0
tensorflow-estimator=1.15.1=pyh2649769_0
termcolor=1.1.0=py_2
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tornado=6.0.4=py36h8c4c3a4_1
urllib3=1.25.10=py_0
werkzeug=0.16.1=py_0
wheel=0.35.1=py_0
wrapt=1.12.1=py36h8c4c3a4_1
x264=1!152.20180806=h14c3975_0
xz=5.2.5=h7b6447c_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
